separate student components
